Monitor/requirements.txt
paho-mqtt~=2.0.0